+++ lib/main.dart
@@ -1,8 +1,8 @@
 import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:sight_calibrator/data/app_database.dart';
 import 'package:sight_calibrator/main_activity.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 import 'app_settings.dart';
 
@@ -41,6 +41,7 @@ class MyApp extends StatelessWidget {
             home: MainActivity(db: db, camera: camera),
             localizationsDelegates: AppLocalizations.localizationsDelegates,
             supportedLocales: AppLocalizations.supportedLocales,
+            locale: appSettings.locale,
           );
         });
   }