+++ lib/settings_fragment.dart
@@ -15,6 +15,21 @@ class _SettingsFragmentState extends State<SettingsFragment> {
 
   @override
   Widget build(BuildContext context) {
+    List<DropdownMenuItem> languages = <DropdownMenuItem>[];
+    languages.add(DropdownMenuItem(
+        value: null, child: Text(AppLocalizations.of(context)!.system)));
+    for (Locale locale in AppLocalizations.supportedLocales) {
+      var languageName = locale.languageCode;
+      switch (locale.languageCode) {
+        case "en":
+          languageName = "English";
+          break;
+        case "pl":
+          languageName = "Polski";
+          break;
+      }
+      languages.add(DropdownMenuItem(value: locale, child: Text(languageName)));
+    }
     return Scaffold(
       appBar: AppBar(
         title: Text(AppLocalizations.of(context)!.settings),
@@ -51,6 +66,26 @@ class _SettingsFragmentState extends State<SettingsFragment> {
                     );
                     setState(() {
                       _appSettings.themeMode = themeMode ?? ThemeMode.system;
+                      _appSettings.save();
+                    });
+                  }),
+              const SizedBox(height: 6),
+              DropdownButtonFormField(
+                  icon: const Icon(Icons.arrow_drop_down_rounded),
+                  decoration: InputDecoration(
+                      labelText: AppLocalizations.of(context)!.language,
+                      border: const OutlineInputBorder(),
+                      helperText: ""),
+                  value: _appSettings.locale,
+                  items: languages,
+                  onChanged: (locale) async {
+                    ScaffoldMessenger.of(context).showSnackBar(
+                      SnackBar(
+                          content: Text(AppLocalizations.of(context)!
+                              .settingsHasBeenChanged)),
+                    );
+                    setState(() {
+                      _appSettings.locale = locale;
                       _appSettings.save();
                     });
                   })